requirements.txt
Django==3.1.1  
django-cors-headers==3.5.0  
djangorestframework==3.11.1
Pillow==7.2.0
pylint==2.6.0
